delete fixed

+++ appli/app/routes/suppression.py
@@ -3,18 +3,21 @@ from flask import render_template, request, redirect, url_for, flash
 from ..models.data import Maisons, Personnes
 from ..utils.transformations import  clean_arg
 
-#la route qui supprime l
+#la route qui supprime la personne et la maison
 
 @app.route("/suppression/maisons/<string:nom_maison>", methods=['GET', 'POST'])
 def sup_maison(nom_maison):
 
     try:
         if request.method == 'POST':
-            nom_maison_a_supp = request.form.get('maison_a_supp')
-            maison_a_supp = Maisons.query.filter(denomination=nom_maison_a_supp)
-            personne_a_supp = Personnes.query.filter(idWikidata=maison_a_supp.idWikidata)
+            #nom_maison_a_supp = request.form.get(nom_maison)
+            maison_a_supp = Maisons.query.filter(Maisons.denomination == nom_maison).first()
             db.session.delete(maison_a_supp)
-            db.session.delete(personne_a_supp)
+
+            if maison_a_supp.idWikidata:
+                personne_a_supp = Personnes.query.filter(Personnes.idWikidata == maison_a_supp.idWikidata).first()
+                db.session.delete(personne_a_supp)
+            
             db.session.commit()
             flash('La maison a été supprimée avec succès.', 'success')
